ARM: dts: meson: atv1200: add the stdout-path to allow enabling earlycon
Support for Meson6 SoCs is currently very limited. It's often unclear
why such a device does not boot. To debug this further earlycon can be
used - however, the atv1200 board did now allow this so far because no
stdout-path was specified (so earlycon did not know which serial console
to use).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>