phy: amlogic: phy-meson-gxl-usb2: support the clock and reset line
The OTG capable USB2 PHY has a reset line and a clock (which are both
shared with other components, such as the USB3 PHY for example).
Extend the driver so it handles this optional clock and shared reset
line. We only trigger the reset during the .init phase since it's a
shared reset line, so triggering it during the driver's .reset
implementation would effectively also only trigger it once anyways.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>