dt-bindings: arm: amlogic: Add support for GXL S905W and the P281 board - WiP
S905W SoC is another SoC from the GXL family. It is a cost-reduced
version of the S905X SoC. The following differences are known:
- S905W supports HDMI resolutions up to 4k@30fps (while S905X support
  resolutions up to 4k@60fps)
- the built-in video decoders of S905W decode up to 4k@30fps (while
  S905X supports decoding up to 4k@60fps)
- CPU speed on S905W is limited to 1.2GHz (compared to 1.5GHz on S905X)

The P281 is a development board from Amlogic which uses the S905W SoC.
It uses the same PCB layout as the S905D P231 board.

TODO: pin-compatible???

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>