pinctrl: meson: meson8: rename eth_tx_clk_50m to eth_tx_clk
On Meson8 this can be used as a 50MHz clock generator if the RMII PHY
requires the TX clock to be generated by the MAC.
Meson8m2 also supports RGMII PHYs - it's MAC can generate a TX clock
with 125MHz, 25MHz or 2.5MHz depending on the line speed.

Rename the pin to make it clear that it can provide a TX clock other
than 50MHz on Meson8m2. This pin currently not used by any board.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>