ARM: dts: meson8b: RGMII doesn't use the eth_txd1_1 and eth_txd0_1 pins
The pin groups eth_txd0_0 and eth_txd0_1 are both connected to the
Ethernet controller's TXD0 data line. The only difference is that
eth_txd0_0 is routed to GPIOH_6 while eth_txd0_1 is routed to DIF_2_P.
eth_txd1_0 and eth_txd1_1 are similar: eth_txd1_0 is routed to GPIOH_5
and eth_txd1_1 is routed to DIF_2_N.

The schematics of the Odroid-C1+ show that it uses GPIOH_5 and GPIOH_6
for the TXD0 and TXD1 data lines. Thus we can drop eth_txd0_1 and
eth_txd1_1 (which are routed to DIF_2_P and DIF_2_N) from the RGMII
group.

Fixes: b96446541d8390 ("ARM: dts: meson8b: extend ethernet controller description")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>