power: reset: gpio-poweroff: add support for multiple GPIOs
Some systems need to trigger two (or more) GPIOs to shut down the
system. In case of the EC-100 there are two GPIOs which need to be
pulled LOW to power off the system.
Extend the driver so it can manage more than one GPIO. All existing GPIO
operations are converted to treat all GPIOs from gpio_descs identically
(this means: all GPIOs are configured as output at the same time and
gpiod_set_value is called for all GPIOs at the same time).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>