clk: meson: meson8b: register the clock controller early
Until now only the reset controller (part of the clock controller
register space) was registered early in the boot process, while the
clock controller itself was registered later on.
However, some parts of the SoC are initialized early in the boot process,
such as the SRAM and the TWD timer. The bootloader already enables these
clocks so we didn't see any issues so far.

Register the clock controller early so other drivers (such as the SRAM
and TWD timer) can use the clocks early in the boot process. Now that we
don't have a "split" initialization process anymore we can also switch
from CLK_OF_DECLARE_DRIVER (which allows registering a part of the clock
controller early, an another part as platform device) to CLK_OF_DECLARE
(which only supports early initialization).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>