ARM: dts: meson8: add the Cortex-A9 global timer
The Meson8 and Meson8m2 SoCs are using four Cortex-A9 cores. These come
with an ARM global timer.
This adds the Cortex-A9 global timer but keeps it disabled for now. The
timer is clocked by the "CPU_DIV4" clock which can change during runtime
(when changing the frequency of the CPU clock). Unfortunately the
arm_global_timer driver does not handle changes to the clock rate yet.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>