Bluetooth: Add new serdev based driver for 3-Wire attached controllers
This is a from scratch written driver to run H:5 aka 3-Wire on serdev based
system with a Bluetooth controller attached via an UART. It is currently
tested on RPi3. It works only without Broadcom integration and that means
firmware loading is not supported. It is easy to add, but in that case
some packet errors happen that are not yet debugged.

It does support data integrity check, but it does not actually support
the sliding window. It is currently limit to 1 packet and on the
receiver side there is no packet loss detection.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>