clk: core: send POST_RATE_CHANGE in clk_change_rate when reparenting
When calling clk_set_rate() we could run into a case where the
PRE_RATE_CHANGE event was sent without a corresponding POST_RATE_CHANGE
event.

clk_core_set_rate_nolock() sends the PRE_RATE_CHANGE event before
calling clk_change_rate(). Inside clk_change_rate() there is a special
code-path for the case where we need to set a new parent clock. If we
only set a new parent but cannot set a new rate for the clock itself
(for example because we're changing the rate of a mux by setting a new
parent) the POST_RATE_CHANGE event was not sent. This was due to the
condition which checks when this event should be sent as it only checked
for a rate change. However, we could have also been modifying the parent
clock so we sent a PRE_RATE_CHANGE event, which means that we are
supposed to send a corresponding POST_RATE_CHANGE event as well.

This was found while implementing CPU frequency scaling for the Meson8,
Meson8b and Meson8m2 SoCs. On these SoCs we need to run the CPU off the
XTAL clock while changing the sys_pll or any of it's children (otherwise
the whole chip may hang). For that reason we're registering a notifier
on the "cpu_scale_out_sel" mux which can switch between the sys_pll (and
a few sys_pll dividers) and XTAL whose purpose is to switch the final
output mux (called "cpu_clk") to XTAL on PRE_RATE_CHANGE and back to
"cpu_scale_out_sel" on POST_RATE_CHANGE. Prior to this patch the
notifier has only seen the PRE_RATE_CHANGE event if all parents of the
"cpu_scale_out_sel" were marked as read-only.

Fixes: 71472c0c06cf9a ("clk: add support for clock reparent on set_rate")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>