ARM: dts: meson8b: add the canvas module
Add the canvas module to Meson8b because it's required for the VPU
(video output) and video decoders.
Add the DMC bus as well which holds the canvas module (on Meson8m2 it's
located in the DMC bus as well, whereas on Meson8 it's located in the
so-called "MMC bus").

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>