pwm: meson: don't use clk_get_rate with the spinlock held
The pwm-meson driver uses a spinlock to protect write access to the
REG_MISC_AB register which is shared between the two channels the PWM
controller.

meson_pwm_calc() uses clk_get_rate() which is using a mutex internally.
With the spin lock held by meson_pwm_apply() we can end up in a "BUG:
scheduling while atomic" error.

Move the spin lock to the only functions which are writing to the
REG_MISC_AB register to get the clk_get_rate() call outside of the
atomic context.

The original "scheduling while atomic" error was observed on a Meson8b
EC-100 board which uses a PWM-based regulator for the VCCK (CPU voltage)
regulator.

Fixes: 211ed630753d2f ("pwm: Add support for Meson PWM Controller")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>