pwm: meson: don't store the polarity as bit-mask in struct meson_pwm
meson_pwm_calc() calculates the PWM register settings (pre_div, hi and
lo) from a list of parameters. All parameters except the polarity are
passed directly to this function.

Pass the pwm_polarity as parameter to meson_pwm_calc(). This makes it
consistent with the other parameters (duty cycle and period).
Additionally it removes a write to the per-controller (not per-channel)
struct meson_pwm which must be protected using the spin lock.

No functional changes intended.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>