mtd: rawnand: meson: fix info buffer size in meson_nfc_read_buf()
Most callers of meson_nfc_dma_buffer_setup() pass an info buffer size of
"nand->ecc.steps * PER_INFO_BYTE". There are two exceptions to this:
- meson_nfc_read_buf() passes only PER_INFO_BYTE (no
  multiplication with ECC step size)
- meson_nfc_write_buf() passes zero (because the hardware doesn't return
  the info data in that case)

In meson_nfc_read_buf() we not enough memory is allocated for the info
buffer which is filled by the NAND controller hardware. This results in
memory corruption because the hardware writes beyond the end of the
allocated info buffer. This corrupts the memory beyond the allocated
info buffer.

Adjust the code in meson_nfc_read_buf() to use the same info buffer size
as all other callers which actually use a non-zero info buffer.

Fixes: 8fae856c53500a ("mtd: rawnand: meson: add support for Amlogic NAND flash controller")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>