[RFC] pwm: meson: don't disable the PWM if meson_pwm_apply() fails
meson_pwm_apply() must not modify the PWM output if the new state cannot
be applied (the error-cases are implemented in meson_pwm_calc(),
examples are the input clock is not valid or the duty cycle cannot be
generated from the pre-divider).

Call meson_pwm_disable() only after meson_pwm_calc() was successful.
This ensures that whenever meson_pwm_calc() fails that we keep running
with the old PWM settings, compared to having the PWM disabled (in case
it was enabled before).
This potential issue has not been seen in the real world (yet).

RFC: should we move the "disable first" logic to meson_pwm_enable()
instead of expecting the caller do to it?

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>