ARM: dts: meson8m2: mxiii-plus: fix the name of the DCDC2 regulator
The DCDC2 regulator output is actually called "VDD_EE" in various
Meson8m2 board schematics. This matches with what Amlogic names it in
their vendor kernel.
While here, also give the regulator an alias as it supplies the Mali GPU
so a phandle to it will be required later on.

Fixes: 35ee52bea66c74 ("ARM: dts: meson8m2: add support for the Tronsmart MXIII Plus")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>