dt-bindings: pci: add binding for the Lantiq xRX200 PCIe controller
The Lantiq xRX200 SoCs embed a PCIe controller which is based on the
Synopsys DesignWare PCIe core.

The controller itself is part of the "fpi" bus using the "dbi" register
space (which also contains the DesignWare PCIe controller registers) and
an "app" register space which is a custom extension. There is a PCIe PHY
which is not part of the PCIe controller register space - instead it is
located on the RCU bus.

Like various other IP blocks on the xRX200 SoCs the endianness is
configured in a special RCU register.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>