dt-bindings: pci: document the Lantiq VRX200/ARX300 PCIe controller
The Lantiq VRX200 SoCs embeds one PCIe controller which is based on the
Synopsys DesignWare PCIe core. The ARX300 SoC embeds two of the same
PCIe controllers.

The controller itself is part of the "fpi" bus using the "dbi" register
space (which also contains the DesignWare PCIe controller registers) and
an "app" register space which is a custom extension. There is a PCIe PHY
which is not part of the PCIe controller register space - instead it is
located on the RCU bus.

Like various other IP blocks on these SoCs the endianness is configured
in a special RCU register.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>