mips: lantiq: switch to PCI_DRIVERS_GENERIC
The Lantiq PCI-Express controller IP is based on the DesignWare PCIe
controller. The driver for this does not work with the legacy MIPS PCI
framework and needs PCI_DRIVERS_GENERIC instead. Also the updated PCI
driver uses the "standard" PCI API instead of the legacy MIPS one.

Select PCI_DRIVERS_GENERIC for the LANTIQ target to disable
PCI_DRIVERS_LEGACY so the new PCI and PCIe drivers can be enabled.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>