drm: meson: get the HHI registers from the "amlogic,hhi-sysctrl" syscon
Add support for parsing the "amlogic,hhi-sysctrl" syscon phandle. This
allows us to get rid of the duplicate ioremap (which is once done in the
syscon driver and then also in the VPU driver).
Fall back to parsing the "hhi" registers if the "amlogic,hhi-sysctrl"
property is absent to stay backwards compatible with old .dtbs.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>