ARM: dts: meson8: fix the 182.1MHz OPP for the Mali-450 GPU
The clock setup on Meson8 cannot achieve a Mali frequency of exactly
182.15MHz. The vendor driver uses "FCLK_DIV7 / 2" for this frequency,
which translates to 2550MHz / 7 / 2 = 182142857Hz.
Update the GPU operating point to that specific frequency to fix
potential devfreq errors (due to devfreq mandating that the rate after
updating the clock output is exactly the rate as specified in the OPP
table).

Fixes: 7d3f6b536e72c9 ("ARM: dts: meson8: add the Mali-450 MP6 GPU")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>