clk: meson: meson8b: make the CCF use the glitch-free "mali" mux
The Mali clock tree should not be updated while the clock is running.
Enforce this by setting CLK_SET_RATE_GATE on the muxes and dividers in
the "mali_0" and "mali_1" clock trees. This makes the CCF switch to the
"mali_1" tree when "mali_0" is currently active and vice versa which is
exactly what the vendor driver does when updating the frequency of the
mali clock.

This fixes a potential hang when changing the GPU frequency at runtime.

Fixes: 74e1f2521f16ff ("clk: meson: meson8b: add the GPU clock tree")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>