ARM: dts: meson8: add the thermal-sensor with the thermal-zones
Meson8 SoCs can use ADC channel 8 to read the chip temperature. Wrap
this in a "generic-adc-thermal" device so it can be used to get the
temperature for the thermal-zones.

The vendor kernel uses the following thermal-zone settings:
<= 70°C:
- CPU frequency limited to 1.608GHz
- GPU limited to 511MHz and 5 cores (pixel processors)

<= 80°C:
- CPU frequency limited to 1.2GHz
- GPU limited to 435MHz and 4 cores (pixel processors)

<= 90°C:
- CPU frequency limited to 0.804GHz
- GPU limited to 328MHz and 3 cores (pixel processors)

Define a thermal-zone for with the same CPU and GPU frequency limits as
the vendor kernel. For now the GPU cores (pixel processors) are omitted
because there's no way to configure them (yet).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>