ASoC: cs4334: Add a driver for the Cirrus CS4334 Stereo I2S DAC
The Cirrus CS4334 is a simple Stereo I2S D/A converter. It does not have
any management interface (like I2C, SPI, GPIOs, ...).
It suports for 16 to 24 bit little endian input formats and sample rates
between 8kHz to 96kHz.

The MXT8234 is a functionally compatible replacement for the CS4334 from
Beijing Times Minxin Technology Co., Ltd.

Datasheets for both implementations are available on the web:
- https://statics.cirrus.com/pubs/proDatasheet/CS4334-5-8-9_F7.pdf
- http://www.mxtronics.com/n107/n124/n181/n184/c697/attr/2625.pdf
  (Chinese only)

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>