ARM: dts: mxiii: add support for the audio outputs
The MXIII and MXIII-Plus boards come with SPDIF and I2S Stereo DAC
(provided by a Beijing Times Minxin Technology Co., Ltd. MXT8234) audio
outputs. The latter is not supported yet (as we are missing a driver for
it). Also the I2S output can be routed to the HDMI TX controller.

Add the SPDIF output as well as HDMI TX audio routing on these boards.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>