pinctrl: meson: meson8b: fix support for GPIO_TEST_N
GPIO_TEST_N is special because it is not part of the documentation in
the public S805 datasheet. There is no documented way to switch it from
an output to an input. The vendor kernel uses the output register to
read the pin value back. The GPIO interrupt controller does not seem to
have a bit for this pad.

The vendor kernel uses AO_GPIO_O_EN_N[31] for the GPIO output value
(setting the bit means HIGH, clearing it means LOW).
The bias can be configured using AO_RTI_PULL_UP_REG[14] (to set the bias
to pull-up or pull-down) and AO_RTI_PULL_UP_REG[30] (to enable the
pull-up/down).
Set the irq, dir and in registers and bits to -1 to mark them as
unsupported and use the pullen, pull, out and in registers and bits from
the vendor kernel.

Additionally the vendor kernel sets bit AO_SECURE_REG0[0] (0xda004000)
with the comment "out put enable" (it does the same for the other
special GPIO: GPIO_BSD_EN). This bit is set on all of my boards (it may
be set by the boot ROM or the vendor u-boot). Fetch the secbus2 regmap
from the "amlogic,secbus2" property and use it to manage the
input/output status of this pin.

Fixes: 0fefcb6876d0d6 ("pinctrl: Add support for Meson8b")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>