drm/meson: Read the CVBS VDAC calibration data from eFuse
Meson8/Meson8b/Meson8m2 SoCs have calibration data for the VDAC (used
for the CVBS output) programmed into the eFuse during the manufacturing
process. Read these values and use them when initializing the VDAC
register to get a video signal out of the CVBS connector.

To keep the code simple, we use a similar code-path for all SoC
generations and default to 0x0 on GXBB and newer - which don't need such
calibration data from eFuse.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>