ath5k: add support for reading the EEPROM data from an nvmem-cell
Some devices supported by OpenWrt don't have an EEPROM connected to the
ath5k chip itself. Instead the EEPROM data for the wifi chip is stored
on a read-only partition of the main flash storage of these devices. Add
support for reading the EEPROM data through the NVMEM framework on these
boards (when configured via device-tree).

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>