ath5k: add support for byte-swapping the EEPROM contents
Some boards with the EEPROM stored in the main flash (which are then
passing the link to the EEPROM data to ath5k using an nvmem-cell) have
the data byte-swapped. In this case the magic word (0x5aa5) is
byte-swapped as well (meaning it reads as 0xa55a).

Read the EEPROM magic word during initialization and decide if swab16 is
needed when reading the EEPROM data. Handle the swapping transparently
in ath5k_hw_nvram_read() so the callers don't have to be updated.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>