dt-bindings: net: wireless: ath5k: add support for cards without EEPROM
Some ath5k cards don't have their own EEPROM connected to them. This is
mostly true for embedded devices which are for example supported by
OpenWrt. In these cases the ath5k EEPROM data is stored in the main
flash, sometimes even as part of u-boot.

Add support for a property which indicates when no EEPROM is connected,
similar to what is in place for ath9k. Also add nvmem-cells for fetching
the EEPROM data as well as the MAC address.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>